Make quit button a no-op.
Calling 'finish' does not seem to work right. For example when the 'new
workout' was started and finish is called in the main activity, the main
activity exits but the 'new workout' activity is still running. We will
need to have a look into that to find a better solution.