Use Exercise class to pass exercises through an Intent.
Rather than just returning the exercise name we now pass the full
Exercise object after adding an exercise. This makes sense as we can
fully set up and customize the exercise from the AddExercise activity
and only have to pass one object back to the caller.

Supposedly using Parcelable would be more performant compared to going
with the default Serializable as we do now. It requires more overhead in
the implementation though and we can keep it in mind if we run into
performance issues.