Add functionality to write a new Exercise to file.
This is a first attempt at saving Exercises on file. We now write the
created as json to the internal storage on the device.
For now it is not read again and a new exercise overrides the old one,
but it is a first step on which we build further functionality.