Add functionality to read Exercise from file at startup.
We now read and show the saved exercise when the app starts. For this we
add a function for the json utilities that parses the given file for an
exercise.

This is only a first attempt to make sure that reading and writing the
file actually works. We will improve this and build on this to at some
point handle multiple exercises in the app.