Use less parameters for call to read Exercises from file.
There is no reason for the main Activity to have to provide the filename,
as we know it and at this point it does not change.
If we have to be more flexible at some point we can always extend the
interface. For now it is better to keep everything simple rather than
trying to accommodate for complexity for which we do not even know yet
whether it is necessary.