Read and write a list of Exercises.
Rather than just creating and writing a single exercise, we now have a
list of exercises that the user can create. We simply write it as a JSON
array to file and read it from there as well at startup.

So far we only print the exercise names at startup. We have to add some
more functionality to actually print them as a proper list on the main
screen. We probably need a function to update the display as well as
right now when an exercise is added, it is only shown the next time you
hard restart the app.