Base ViewHolder on FrameLayout rather than View.
The tutorials are somewhat confusing and buggy in this detail. If we
strictly implement them (as we did) they actually don't work. We now
simply roll with the flow and actually use the FrameLayout that we have
specified to show our view. Now the different exercises are actually
listed on the main screen.