Add initial support for exercise context menu.
We now show a context menu on long press in the main view. As per
recommendation on the Android guidelines it is a ActionMode [0] menu.
This will at some point allow us to do batch selections as well.
As it is clicking to delete the exercise does not have any effect yet
though.

[0] https://developer.android.com/guide/topics/ui/menus.html#context-menu