Add functionality to delete exercises from list.
We now can delete exercises that are long pressed on in the list of
exercises. It is the first iteration on doing this but it seems to work
alright. One issue to improve is the updating of our json backend.
Rather than manually writing the exercises to file again we should
probably introduce a class that provides an update method.

Another cause for concern may be our use of static member variables to
allow access to for example the application context. At this point I'm
not sure whether this is a bad approach or how to do it better.