Rename ExerciseAdapter to ExerciseViewAdapter.
ExerciseAdapter is somewhat ambiguous as it does not make clear what the
adapter is meant to be used for. Especially now as we want to have a
kind of container class for exercises it makes sense to have explicit
and descriptive names for the classes (whilst at the same time trying
not to use hugely long names).