Update exercises on file after removing or adding exercises.
This is where it starts to make sense to have a wrapper class. Rather
than having to manually update the exercises every time we are modifying
them in one of the Activities, we simply make it part of the process of
modifying them.

And, if in future, we want to optimise the process and for example only
make batch changes, we can easily change that in the wrapper class
without having to touch all other instances where it is used.