Add initial attempt at class for exercise session.
The Session class is our first attempt at modeling an exercise session.
With it we will store a number of sets (each with repetitions and
weights) for a given date.

It would have seemed like a good idea to make Session an inner class of
Exercise, as it is only used there. According to the Java docs [0]
though it is a bad idea to use the Serialisable interface with inner
classes. For this reason we keep it as a separate class and use it like
this.

[0] http://docs.oracle.com/javase/tutorial/java/javaOO/nested.html