Add initial ViewExerciseActivity.
Short clicking on one of the exercises in the main activity now opens
the 'view exercise' activity. For now it is empty, but we will use it to
show a list of all sessions for that exercise.

One issue that is apparent though is that when returning from the
activity to the main view, the original view is reset and the view is
placed on the first item in the list again. This is something we should
fix.