Turn the ExerciseWrapper into a true singleton object.
Whether using singleton objects is a good idea seems to be base of a lot
of discussions. I think it makes sense in our case where we want to load
the Json files only once and have them only once in memory too.

We can always revisit this later on if we come across issues.