Add first attempt towards filling Session list in Exercise view.
In theory we should be able to use 3 TextViews next to each other to
show the date, repetitions and weights for each session. This commit
adds basic support for this. Half the values are hardcoded as we can not
access them from the Session class yet, but we will fix this in future
commits. Also, the actual text views are right next to each other make
the text somewhat unreadable which needs to be improved.

Basically this is a first draft or prove of concept on which we can
base further development.