Only use one set per Session.
Rather than having separate arrays in a Session object containing all
the sets that we did for that exercise, we now simply have one set per
session. This makes it a lot easier to access specific sets, and to
extend the information they provide. This way we can for example easily
add a comment per set without having to jump through a lot of hoops.

We still know which sets were done at the same day and in which order,
as we can deduct that information from the Date attribute when the
session was created.