Add title for Session view.
We now show 'Date Weight Repetition(s)' as a title above the actual
sets. It does not quite match the layout and design yet, but we will
fix this in future commits.

We also hardcode the strings in the layout xml, as it might be better
for performance and maintainability. We are not going to modify them
during runtime anyway.