Add blank NewSessionActivity.
For now it is just a blank Activity, but we will use it to add a Session
to an Exercise.

Note that we set the launchMode of ViewExerciseActivity to 'singleTop'.
This means that when pressing the up or home button, rather than
creating a new Activity, the old active parent is pushed to the top.
This avoids a crash as otherwise the provided Intent does not contain
the exercise position that is meant to be displayed. Also, it would be
wasteful to create a new Activity when we do actually want to return to
the previous one.

More details on the launchMode can be found here:
https://developer.android.com/guide/topics/manifest/activity-element.html#lmode