Abort ViewExerciseActivity if it is started without correct Intent.
Calling finish alone does not exit the Activity. We have to call
finish and return.

Also document a bit better on why it should not happen, and how it can
happen, and how we try to avoid it. This makes it more clear on why we
have that condition there in the first place.