Set MainActivity launchMode to singleTop.
This way the original position in the list of Exercises will be restored
when the 'up button' is pressed from one of the child Activities.
Basically this way the 'up button' behaves the same way as when pressing
back. I think this makes sense as going back is usually the intended
action when going out of for example the ViewExerciseActivity.