Return empty or partial list rather than null on error in JsonUtils.
The app currently crashes when there is no existing exercise data on the
device, as the exercise array is set to null and then accessed.

Returning an empty list or partially filled list from JsonUtils fixes
this issue. It might be worth considering whether it is better to return
an empty list or a partially filled one, though. For now it will be fine
though.