Only allocate memory once when loading exercises.
We know how many exercises we have read from the json file, and it makes
sense to allocate the array accordingly. Note that the return type of
readExercisesFromFile is still List<Exercise> and we only use the
ArrayList specific ensureCapacity method inside of
readExercisesFromFile. This way, if we change the underlying list
implementation, we only have to update it in readExercisesFromFile as
all callers will still be able to refer to the List interface.