Return position on which new exercises are added.
It can be useful for the caller to know where a new exercise was added.
For now we know that it will simply be the end of the list, but at some
point we may have sorted Exercises and it will not be as apparent.

Returning the position from where the exercise is added makes it future
proof as it we then only have to update the ExerciseWrapper class.