Notify about inserted items specifically.
According to the documentation for the RecyclerView it is better to
notify about changes specifically and use notifyDataSetChanged only as a
last resort. The reason being that it causes any adapters using the
items to assume that the whole list may have changed, and it will be
completely reloaded. By notifying only about the specific inserted item,
the adapter only has to load the affected item.