Add number pickers for weight and repetitions to new session activity.
We now have to scrolling number pickers to select weight and repetitions
for new workout sessions. They are quite bare as they are, not spaced
correctly yet and the picked values are not used in any meaningful way,
but it is a start to continue working on them.

With these we basically have buttons for the minimum required inputs to
be able to add a new session, and we only have to place and style them
right and actually use them.