Place all Json output in JSONObject.
I think this may be a better approach towards (de)serialising all data
for the app. Because even though right now we only have a list of
Exercises, we will add more fields in the future. And having a
JSONObject wrapped around everything means that we can simply grab
future fields out of it. This should make it a bit more maintainable and
future proof.