Add support for writing and restoring Sessions from Json.
The app now actually saves and restores the Session data for each
Exercise.

Unfortunately the JsonUtils code is getting a bit unwieldy. I'm not
quite sure yet how to improve it, but we may consider adding a couple
more methods to make it more readable. For now it does what it is
supposed to do though.

We may as well consider to store the Session dates as a String rather
than a long value. It would make the actual Json file more readable for
a human, but would mean more effort in parsing the date in the code. For
now using longs is the simplest approach code-wise.