Write changes to file when Sessions are added.
As we are not modifying the Exercises through the ExerciseWrapper, we
have to explicitly ask it to update the Exercises on file. For this we
simply add a static method that can be called from anywhere in the
project.

Ideally any changes would automatically be synced, without having the
caller to be aware of methods that do and don't take care of this. I'm
not quite sure how to achieve this at this point, though. For now it is
sufficient to do it explicitly and have documentation for the methods
that mentions when changes are being applied.