Initialise New Session with previous weight and repetitions.
We now initialise the New Session dialog with the previous values for
weight and repetitions. This makes sense, as during a workout the
individual Sessions or sets are often close to each other. This way
there are less adjustments necessary by the user when a new Session is
added, which is always a good thing.

It feels a bit clunky to have to provide the last values from the
Exercise object through the Intent into the NewSessionActivity. I'm not
quite sure how to simplify it though as the previous values are
different for each Exercise and the Session itself is not aware of the
other Session objects. We might keep it in mind, and leave it for now as
it seems to work well.