Add new Sessions via Exercise object.
Rather than touching the Sessions list directly, we now add Sessions
through the Exercise object. This way we make sure that for example the
last used weight and repetitions are updated for when we next add a
Session.

This also means that the information about where the Session was
inserted is owned by the Exercise again, which is good. Because if we
change anything about this, we will only have to change it in the
Exercise class itself. It makes sense for the ViewExerciseActivity to
not be aware of these details.