Remove mSessions member variable from ViewExerciseActivity.
We do not want to modify the list of Sessions directly from the
ViewExerciseActivity. It makes more sense to make use of the methods
provided by the Exercise class to do so, as it will for example take
care of updating what the last used weights and repetitions were.

This is slightly different from the MainActivity class where the data
set is actually just a wrapper (the ExerciseWrapper to be precise) that
takes care of the logic around adding and removing Exercises. We don't
want the logic for adding and removing Sessions to be part of the
ViewExerciseActivity as it does not belong there. Its only purpose is to
show an Exercises. Removing the Sessions list as a member variable forces
us to use the mExercise variable and go through the Exercise object instead.