Do not sync changes to file again when loading via Json.
Writing the changes to file again whilst we are loading them from there
does not make sense. Though to be accurate the
ExerciseWrapper.notifyExercisesChanged does not actually write the
Exercises back to file, as the ExerciseWrapper itself is not initialised
yet at that point. But, this is more of a side effect of the code being
the way it is. It is probably a better approach to be explicit about it,
and show via parameter whether the changes are meant to be synced or
not. It is slightly more verbose but the explicitness makes up for it in
my opinion.