Use the selected date when adding a new Session.
We now actually use the date that the user picked when adding a new
Session. We simply default to the current date if there was no
selection, and otherwise go with the selection. Note that the
implementation is not ideal yet. The Session is not inserted in the
right slot if the date is before an existing one. Also, there is no text
on the screen to show the user which date was selected.

We should also consider whether it would make sense to open the
DatePicker dialog with the Date that was previously used. Right now it
always starts with the current date.