Add debug print for when notifyExercisesChanged is called uninitialised.
Ideally we do not want to have it called if it is not necessary, for
example when the existing Exercises have not been initialised yet. It
does not happen the way it is in the code, but a debug print will make
it more clear if it happens in the future.