Increase margins in Session view to improve readability.
They are not perfect yet as they are, but a bit more space does make it
more readable. We also remove the card_testview_margin as there is no
point in defining it if we can simply use the pre-defined Android
margin. Hopefully this will make it easier to use it for differently
sized devices, as the same margin may be differently sized.