Sort Sessions when a new one is added.
Now new Sessions are correctly inserted in order when an earlier date is
selected than the current one.

Note that this implementation is quite inefficient. We should be using a
SortedList [0] to take care of the ordering and sorting. It would
provide callbacks for the RecyclerView as well. I am not quite sure how
to integrate the Android support library though so I will leave it for
improvement in the future. Also, we do not have explicit ordering for
Sessions on the same date yet. Right now it simply orders them
chronologically in the order they are added, but we should consider
improving this in the future.

[0] https://developer.android.com/reference/android/support/v7/util/SortedList.html