Show selected date in new Session dialog.
We now show the date that is going to be used for the Session to be
added in the dialog. This is good, as only having the button makes it
seem like a date has to be picked before a session is added.

This also is a good opportunity to work in a cleaner interface between
the DatePickerFragment and the NewSessionActivity. As per the
developers' guide [0] the NewSessionActivity is now notified via
callback when the user picks a date. This allows us to remove the
somewhat awkward handling of the date through the DatePickerFragment
that we had before.

[0] http://developer.android.com/training/basics/fragments/communicating.html