Replace System.out.print with Log calls.
Using System.out for printing debug information is not recommended as it
is quite unflexible. The Log module is better as it allows to use
different levels of logging that can be blocked or filtered for. It also
includes the TAG for the respective application so that it is clear in
the Android log which application printed the message.

We were already using Log calls in most of the app, and this commit
simply changes the remaining two instances.