Store versionCode on file.
The versionCode is the identifier we use to determine what version of
the app is in use. As we may change the format or contents of the json
files in future, it makes sense to have a reference to which version was
used to store the respective data. This way any future version of the
app can handle the previous version's data accordingly.

For now we just write the version and read it again, without any special
handling as we haven't released the app yet and no special consideration
has to be taken into account.

Also we move the handling of writing the versionCode into its own
method, which makes the overall code structure a bit cleaner. We should
consider doing this for the reading and writing of exercises too.