Adjust Session description margins to better match the content.
It is a bit of a messy approach, but the Date, Weight and Repetitions
descriptions now sit a bit better aligned above the listed actual
exercise sets. I am not quite sure what a good way is to do this right,
as somewhat randomly trying different margins until it works does not
seem to be a sensible approach in solving this. It works fine for now,
but in future we should find a better way to solve this issue.