Align weights and repetitions in exercise view.
We now right-align the values for the weights and repetitions. It looks
a lot cleaner that way when there are values of different lengths.
For example 5, 15 and 115 now show as
  5
 15
115
We have to specify both gravity and textAlignment as we support API
versions older than 17 which do not use textAlignment yet.
We also shift the Weight description in the exercise view slightly to
align with the actual values below.