Align date text with DatePicker button in NewSessionActivity.
It looks cleaner when the text for the picked date is vertically aligned
with the button to pick the date. We simply achieve this by using an
explicit size for the buttons in the Activity, and using the same size
for the date text. There may be a better approach in solving this, but
for now this works fine.