Use double instead of int for Session weights.
Interestingly enough it did not occur to me during the design, but it is
quite common to use weights with decimal values. A lot of weight
machines for example have half weights that can be added. To make this
possible we now use double for all weight values in the app.

Note that it is not yet actually possible to add a session with a
decimal value weight. We will have to update the NumberPicker for this.