Add second NumberPicker to select decimal for new Session weight.
We now use a second NumberPicker for the user to select the fraction
part of the weight for a Session. Using two NumberPickers instead of
just one with double values makes sense because it is easier to the user
to select. Having full double values on one number picker would result in
an awful lot of values to scroll through.

Note that the splitting of the double weight into whole and fraction is
a little bit dodgy and only working for one digit in the fraction part.
We raise an Exception though if we should provide longer fractions in
the future, which hopefully avoids any maintenance issues.
For now this approach works fine, and we may consider a more elegant way
of doing it in the future.