Scroll view to newly added Exercises and Sessions.
Now the newly added items are actually visible when they are being
added. Before they would be added but the user had to manually scroll to
see them.

Note that adding a new Exercise does not scroll smoothly when the
keyboard is still in view when the Exercise is added. We may have to
have a look into that at some point, but for now it is fine as it is.