Show confirmation dialog before deleting an Exercise.
At this point Exercises can not be recovered once deleted. For this
reason it makes sense to show a dialog and to ask the user whether they
actually want to delete it.

It comes with a fair bit of code but works as expected and actually
looks fine as it is.

Note that the ActionMode that selected the Exercise actually finishes
before the user confirms or cancels the deletion dialog popup. It is
important here to not reset the mSelectedPosition variable before any of
the actual dialog callbacks are called.