Focus name and bring up keyboard when NewExerciseActivity is called.
At the moment the only action when adding a new Exercise is to specify a
name. Rather than the user having to click on the input field, we now
bring up the keyboard automatically when it is started.

Note that it does not seem to be necessary to request the focus for the
EditText, but I reckon it is better to be more explicit here. Also, the
automatic focus may change once we add more elements to the Activity.