Print Session date only for topmost Session per day.
We now only print the date if it is the topmost Session overall, or the
topmost Session on a given day. This makes the Session overview a bit
more readable, and makes it more obvious which Sessions happened on
which day.

We set the width of the date field to a fixed size so that the empty
entries do not affect the placement of the weight and repetition
fields.