Move onSameDay Date function into utility class.
There are a couple of spots where we have to take apart Dates by hand.
To avoid code duplication it makes sense to move these into a utility
class. For this reason we add Util.java. There will probably be further
functions to be added, but for now we just have onSameDay to determine
whether two given Dates are on the same day.