Only set lastWeight and lastRepetitions for most recent Session.
We sort the Sessions in descending order, but load them chronologically
at startup of the app. This means that the latest added Session after
startup is always the oldest one. Only setting the lastWeight and
lastRepetitions values when it is the newest Session or on the same day
as the newest Session(s) avoids this issue.