Mark and unmark a long-clicked Exercise.
This has been in our TODO for quite a while. We now support
functionality to show that an Exercise has been selected to be acted on
through the ActionBar. We reset the selection again when the action is
cancelled or a specific action is selected.

For now we only show "(selected)" behind the Exercise name, but we
should think about a more prominent display for this.