Add initial support for showing the number of days since the last Session.
We now show behind each Exercise how many days have passed since it was
last done. Ideally using the String in the resources file makes it
easier to translate it if necessary in the future. I am not quite sure
though how for example different sentence structure would be taken care
of.

Note that showing this information behind the Exercise name is not ideal
and should be moved. Also, the information is only updated at startup
and not when new Sessions are added. We may also consider whether we
want to compute the time every time an Exercise is loaded or whether we
simply want to calculate all entries at startup when the Exercises are
being loaded.