Name String variables for Dates 'dateString'.
We are using a lot of Dates and Strings referring to show the respective
dates. It gets a it confusing when both Strings and Dates are called
'date' interchangeably. To make it more clear we now name all the String
variables that contain dates 'dateString', and reserve the name 'date'
for actual Dates. This should make the code a little bit more readable.