Show elapsed number of days between Sessions.
We now show the number of days that passed between exercise Sessions.
This way it is clear how much time passed between the Sessions without
having to count the days. Unfortunately it takes a bit fair bit of
horizontal space to show the information. Though for now it is okay, and
we can look into alternative presentations once we require more space to
show more information on the Session lines.