Avoid leaking the Activity context.
Leaking the actual Activity as a Context can result in memory leaks [0].
Especially for something like the static ExerciseWrapper this is a
problem. The MainActivity would be recreated after for example an
orientation change, but the ExerciseWrapper would still maintain a
reference to the old Activity and stop it from being garbage collected.
For this reason references to the Activity context should be avoided,
and the application context used instead. Update the usages in our code
to reflect this.

Note that mActivity in ExerciseViewAdapter not can not be replaced by a
Context, as it calls startActionMode which is only available from an
Activity directly.

[0] http://android-developers.blogspot.com.au/2009/01/avoiding-memory-leaks.html