Move logic for time since last and between Sessions into Exercise class.
Rather than having the logic for determining the time between Sessions
and since the last strewn all over the code base, we now have both
inside the actual Exercise class. This has a couple of benefits. It
makes it more clear where functionality to determine the time is being
used, as all occurrences of it make use of the methods in the Exercise
class. Also, it simplifies the code of the Exercise and Session view
adapters, which really should only contain logic for displaying the
items.

One drawback for this approach is that now the Exercise class determines
what for example the SessionViewAdapter shows for each Session. I would
say this is outweighed by the advantage of clearer logic though.