Show number of days since topmost Session in Session list.
So far we only showed the number of days for Sessions that were not at
the top. Now we show it for the topmost Session too. This makes sense
for two reasons. For once it makes the look of the topmost Session match
the ones below which is more pleasant to look at. As well it shows the
time since the most recent Sessions which is not immediately clear if it
was a while back when the user opens the Session view.