Update the Exercises when the app was last started on a previous date.
We now update the list of Exercises when the app is resumed on a new
day since the last time.

Note that it may not quite work as it is as only half a day may have
passed since a specific Exercise, and we do not take this into account
yet in Util.daysBetweenDates.