Remove Context member variable from Exercise class.
We can not have it as a member, as it breaks serialisation of Exercise
objects when we insert Exercises in Intents. Providing the context
through the method call seems cleaner too as an Exercise otherwise does
not care about it at all.

This is a prime example of the use of regression tests. This issue
caused the app to crash when adding Exercises, but it did not happen
until days and changes later as I was not testing this specifically.