Use normalised Dates for 'days since' calculations.
This should avoid the issue where for example 2.5 days have passed
since an exercise but the app only shows 2 because it rounds down. We
want to count it as 3 days as it is 3 days on the calendar. We normalise
the Dates to midnight as the 'Date.time()' in Dates is set as milliseconds
since midnight on Jan 1 1970, so it makes sense to count the days until
midnight again.

This simplifies the calulation in Util.timeSince as we do not have to
take half days into account anymore.

Note that we should still keep an eye on the Dates as I would not be
surprised if there were still corner cases where the calculation does
not work as expected.