Update Session below newly added one if necessary.
If we add a new Session and there is a Session on a different day below
it, we need to update both in the RecyclerView.

It is not overly complex logic so we could have it in the
ViewExerciseActivity class, but it seems sensible to keep it in the
Exercise itself. That way it is all in one location and not strewn
around the code base, and we can find it simply by searching for the
usage of the methods.