Store Exercise json file on external storage.
We now store the Exercises on external storage. This way they are
publicly readable and for example can be accessed via adb.
We also pretty print them so that they can actually be comfortably read.

Note that we do not read the Exercises if write access is not possible.
We may consider changing this in the future, but for now it is easier in
the code to do it this way.