Always request external storage permission.
It looks like I may have misunderstood the documentation when it stated
that from API level 18 upwards on the permission does not need to be
requested for private data. Or maybe we are not using private data. The
way it is, the app is not able to write to storage on the emulator on
API level 23 and the permission is required.

For now we always request it and maybe look into it in the future to
check whether we can avoid requesting it.