Use Material Design through AppCompat.
This is the first attempt at using Material Design for the app. The
colours are more or less randomly picked from the ones available in the
editor that comes with Android Studio. To make it work we need to
specify a later compileSdkVersion (21), but it still works for API 18 on
my actual device.

For now the setup is based on [0] and [1].

[0] http://android-developers.blogspot.com.au/2014/10/appcompat-v21-material-design-for-pre.html
[1] http://developer.android.com/training/appbar/index.html