Move weight-based Exercises into their own class.
The end goal is to be able to have time-based Exercises. As a step
towards that we add a new class for weight-based Exercises that inherits
from the more generic Exercise class. For now we discern the type by a
'type' method. There may be better ways of doing this, but we can work
those out whilst we continue on implementing support for time-based
Exercises.