Make Exercise and ExerciseSession classes abstract.
Using these as abstract classes makes sense, as we do not want them to
be instantiated by themselves. The official documentation [0] has more
information. We still have to do some casting, but it is not too bad. It
also allows us to override for example the 'add' method in case of the
Exercise class.

[0] https://docs.oracle.com/javase/tutorial/java/IandI/abstract.html