Show a hint screen in ViewExerciseActivity when no Sessions have been added yet.
We now show a basic hint to make it clear that an empty exercise is
being shown and how to progress from there. It is not quite perfectly
themed yet, but looks alright for a start. Also, we should consider
crossfading [0] between the empty exercise view and when the first
Session is added as the transition is quite abrupt as it is.

[0] https://developer.android.com/training/animation/crossfade.html