Add Uncrustify config file.
This is a first step towards maintaining a proper Uncrustify [0]
configuration for this project. It will help to avoid unnecessary
formatting errors. The configuration does not cover all cases yet, but
we can extend it when we come across issues. We are checking the changes
on commit via pre-commit hook [1].

[0] http://uncrustify.sourceforge.net/
[1] https://github.com/githubbrowser/Pre-commit-hooks