Update app logo.
The previous logo was just a temporary solution. This is a first attempt
at something original. We now use a stylised barbell as a logo, which
looks somewhat okay, but at definitely better the the previous one.

It is simply exported from Inkscape and imported in Android Studio. We
may consider exporting the different resolutions directly rather than
resizing them from the exported png, as we then can use the lossless
scaling the vector original.