Show decimal point in weight decimal NumberPicker.
We now make it clear that the second NumberPicker for the weight is for
the decimal value by showing a decimal point before the digit.
To make it work with Android we have to provide a String array of
possible values. Unfortunately it would then show a regular keyboard
when selected as well. This is what the workaround with findInput
avoids, which is based on [0] and [1]. This actually works quite well
overall.

[0] http://stackoverflow.com/a/19442674
[1] http://stackoverflow.com/a/37190261