Move generic functionality for new Session activity into its own class.
We are going to have to create Sessions for both weight-based and
time-based exercises. It makes sense to use the same date-picking
functionality for both, so we move this logic into the newly created
NewExerciseSessionActivity class.

Note that we still have to find a good way of initialising the shown
date at the start of the Activity. For now it is in the
NewWeightBasedSessionActivity class, but at the latest when we add
time-based Session we should consider how to solve this through the
generic parent class.