Only add a new Exercise when it is actually supported.
We do not support time-based Exercises yet at this point. So for now
when this type is selected, we simply do not add it.

The calls for raising the Exception are not actually necessary, but I
think they may be helpful if we decide to add additional Exercise types
in the future.