Move functionality to set initial Date text into its own method.
We are going to use it for both time- and weight-based Sessions, so it
makes sense to have it somewhere central; in this case the parent class.

It would be nice to set it as part of the onCreate call, but I do not
see how it would be possible before loading the actual layout. So for
now we do it this way.