Initialise time and distance NumberPickers for new time-based Sessions.
This is a first step towards being able to add new time-based Sessions.
For now we use a very basic layout and just have two NumberPickers to
select time and distance. For now values can be selected but it is not
possible yet to actually create Sessions.

But, we can work on this as a baseline. Ideally we will have the ability
to properly pick times (with hours/minutes/seconds) and distances with
decimal parts.