Return and show time-based Sessions when they are added.
We now actually create and return time-based Sessions. We do not store
them in json yet, but will add support for this in future commits. Note
that we have yet to split the layout for displaying Sessions into time-
and weight-based ones, as now we still show weight and repetitions as
headings rather than duration and distance.

The ViewAdapter that we add is basically the same as for weight-based
Sessions. Even though I did not see a good way of generalising it. It
may not matter as much though as it is pretty much boilerplate only with
not too much specific logic for actual handling of the Sessions
themselves.