Rename package from com.dermobbbda.gymcoachbob to com.baertiger_baer.gymcoachbob.
This is a somewhat big commit as it involves pretty much all files, but
it makes sense. baertiger_baer sounds better as a project name / company
name than dermobbbda. Neither is great, but I can not think of something
better for now. Unfortunately, once we publish the app on Google Play,
we will not be able to change it anymore.