Add wrapper class to disable logging for release builds.
As per the Android guidelines [0] apps should not print any logs. There
is a fair bit of discussion around how to do this best [1] and I think
the wrapper class is the best approach. It does not require any
modifications of the actual log calls and it just adds one extra class
in the sources. One mentioned drawback is that the code still generates
the Strings to be printed but I would say that in our case there should
be no real performance penalties as we a) do not have any heavy logging
and b) do not really have performance intensive operations.

Note that error logging is still enabled as these do not happen often
and are worth mentioning in the log.

[0] https://developer.android.com/studio/publish/preparing.html
[1] http://stackoverflow.com/questions/2018263/how-do-i-enable-disable-log-levels-in-android