Remove unused context menu entries.
We were going to implement support for a settings dialog or exiting the
app. Neither of them actually do anything at the moment. As we are
preparing for a proper release it makes sense to clean this up and get
rid of them. If we decide to introduce them again we can simply undo
these changes.