Hardcode debug print tag in Log wrapper calls.
As all our debug print calls use the same tag there is no point in
specifying it for every single call. We now have it hard-coded in the
wrapper class and do not need to set it for individual debug print
calls. This makes the code a bit more readable and easier to type as
well.