Use locale-based formatting for date Strings.
So far we had hardcoded the formatting for the dates that are being
shown in the app. Now we use a solution based on [0] to actually respect
the locale set on the device to use an appropriate date format.
This way it will look right for users in any country and not just in
Australia. It is a little bit fiddly as we need access to a Context
object but otherwise this seems like a pretty straightforward solution.
And by using the Util class we only have to do it once.

[0] http://stackoverflow.com/a/456941