Add landscape layout for adding weight-based Sessions.
It is not the prettiest layout as it is, but it is functional as a
start. Before the 'add set' button was out of view, and now it is
possible to actually add a Session again. We will want to change this a
bit more in the future and do the same for time-based Sessions.